trantienloi update

+++ demo/demo/src/main/java/com/trantienloi/jobhunter/demo/controller/UserController.java
@@ -2,14 +2,17 @@ package com.trantienloi.jobhunter.demo.controller;
 
 import org.springframework.web.bind.annotation.RestController;
 
+import com.trantienloi.jobhunter.demo.domain.RestResponse;
 import com.trantienloi.jobhunter.demo.domain.User;
 import com.trantienloi.jobhunter.demo.service.UserService;
+import com.trantienloi.jobhunter.demo.service.errors.IdInvalidException;
 
 import java.util.List;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -42,10 +45,24 @@ public class UserController {
         return ResponseEntity.status(HttpStatus.OK).body(loi);
     }
 
+    // local
+    // @ExceptionHandler(value = IdInvalidException.class)
+    // public ResponseEntity<String> handleIdException(IdInvalidException
+    // IdException) {
+    // RestResponse<Object> rs = new RestResponse<>();
+    // return ResponseEntity.badRequest().body(IdException.getMessage());
+    // }
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<String> deleteUserByID(@PathVariable("id") Long id) {
+    public ResponseEntity<RestResponse<String>> deleteUserByID(@PathVariable("id") Long id) throws IdInvalidException {
+        if (id > 1500) {
+            throw new IdInvalidException("id khong lon hon 1500");
+        }
         this.userService.deleteUserByID(id);
-        return ResponseEntity.status(HttpStatus.OK).body("Xóa thành công user với id " + id);
+        RestResponse<String> rs = new RestResponse<>();
+        rs.setData("Xóa thành công user với id " + id);
+        rs.setStatusCode(HttpStatus.OK.value());
+        rs.setMessage("CALL API SUCCESS");
+        return ResponseEntity.status(HttpStatus.OK).body(rs);
 
     }
